requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
neo4j>=5.14.1
python-dotenv>=1.0.0
google-generativeai>=0.3.2
tiktoken>=0.5.2
pydantic>=2.5.2
PyPDF2>=3.0.0
python-magic>=0.4.27
Flask>=2.0.0 